append index unsupport message
